[add]csvファイルの読み込み処理を追加

+++ model/tell_type_model.go
@@ -1,7 +1,7 @@
 package model
 
 type TellType struct {
-	TellTypeId int64 `xorm:"'tell_type_id' pk autoincr" json:"tellTypeId"`
+	TellTypeId int `xorm:"'tell_type_id' pk autoincr" json:"tellTypeId"`
 	TellType string `xorm:"'tell_type' not null" json:"tellType"`
 }
